Reorganize
 ## protocol/malicious

protocol/malicious is no more; this previously contained the malicious
validator.  That has been moved.

 ## protocol/context

protocol/context contains the core traits and implementations for the
module as a whole:

Context is the trait that all contexts implement; this is unchanged.

UpgradableContext is a trait for contexts that can be upgraded.  This
contains associated types for the validator that is produced when
upgrading, plus the type of the context that is produced by the
validator.  It also contains a function that produces the identified
validator.

UpgradedContext is a trait for contexts that have been upgraded.  While
UpgradableContext and Context are not bound to any field type, this is
bound to a type because that is the type of the validator/accumulator
that this contexts uses.  This contains an associated type that
identifies the type of secret sharing that it uses (`Replicated<F>` for
semi-honest, `MaliciousReplicated<F>` for malicious).

SpecialAccessToUpgradedContext is the trait we use for things that some
functions need to do to an upgraded context, but we don't want to see
being used generally.

Base is the semi-honest implementation of Context that we used to have.
No change here except to its name.

 ## protocol/context/prss

This is unchanged.  It contains wrappers so that we can instrument all
of the PRSS functions.

 ## protocol/context/semi_honest

This module contains implementations of semi-honest contests.

Context is a thin wrapper around Base that implements Context and
UpgradableContext.  This context upgrades to Upgraded, via a semi-honest
Validator implementation (see the validator module).

Upgraded is a thin wrapper around Base also that implements Context,
UpgradedContext, and SpecialAccessToUpgradedContext.

 ## protocol/context/malicious

This module contains implementations of malicious contexts.

Context is a thin wrapper around Base that implements Context and
UpgradableContext.  This context upgrades to Upgraded, via a malicious
Validator implementation (see the validator module).

Upgraded is our old MaliciousContext.  It does all of the malicious
protocol functions we need, including functions that allow upgrading of
inputs.

This module used to have a lot of stuff around upgrading things to
malicious.  Those have been moved to the upgrade module.

 ## protocol/context/validator

This contains a new Validator trait that is used in the context upgrade
process.  The context function produces an upgraded context, the
validate function validates the outputs.

The SemiHonest struct implements Validator for semi-honest protocols.
This is nothing but a thin wrapper.

The Malicious struct implements Validator for malicious protocols.  This
is the all-singing version we used to refer to as MaliciousValidator.
It uses the renamed MaliciousAccumulator to do its work.

 ## protocol/context/upgrade

This contains the UpgradeContext and UpgradeToMalicious traits and
implementations that we use to upgrade to malicious.

 ## protocol implementations

The major change here is to unite all protocol implementations.  All
cases where we had x() and malicious_x() have been changed to use the
malicious implementation.  Validators can now be used.  The protocol now
executes based on whether you instantiate SemiHonestContext or
MaliciousContext to start with.

Unfortunately, along with removing a bunch of functions, there are some
downsides:

* The new traits and whatnot might be clearer, but they are a little
  bigger.
* The generic functions are more generic.  That means more generic
  parameters and sometimes more generic constraints.  This is most
  noticeable for the more complex parts of the protocol where the trait
  bounds can be a little complicated.

The net effect is that this hasn't really reduced lines of code overall,
though it should make it easier to implement new protocols, because
there is only one code path to consider.
