+++ src/protocol/basics/mul/sparse.rs
@@ -195,8 +195,7 @@ pub(in crate::protocol) mod test {
         },
         protocol::{
             basics::{mul::sparse::MultiplyWork, MultiplyZeroPositions, SecureMul, ZeroPositions},
-            context::{Context, UpgradableContext, UpgradedContext},
-            malicious::Validator,
+            context::{Context, UpgradableContext, UpgradedContext, Validator},
             BitOpStep, RecordId,
         },
         rand::{thread_rng, Rng},
