+++ src/protocol/sort/secureapplyinv.rs
@@ -30,8 +30,7 @@ mod tests {
         use crate::{
             ff::{Field, Fp31},
             protocol::{
-                context::{Context, SemiHonestContext, UpgradableContext},
-                malicious::Validator,
+                context::{Context, SemiHonestContext, UpgradableContext, Validator},
                 sort::{
                     apply::apply_inv, generate_permutation::shuffle_and_reveal_permutation,
                     secureapplyinv::secureapplyinv_multi,
